+++ ios_audio_test/Graph.swift
@@ -10,9 +10,9 @@ import SwiftUI
 
 struct Graph: UIViewRepresentable {
     
-    @Binding var values: [Double]
-    var minValue: Double = -1.0
-    var maxValue: Double = 1.0
+    @Binding var values: [Float]
+    var minValue: Float = -1.0
+    var maxValue: Float = 1.0
 
     func makeUIView(context: Context) -> GraphView {
         let view = GraphView()
@@ -28,7 +28,7 @@ struct Graph: UIViewRepresentable {
 }
 
 struct Graph_Previews: PreviewProvider {
-    @State static var values: [Double] = [0.0, 0.1, 0.2, 0.3, 0.4, 1.0, 1.0, 0.5, 0.5, 0.0]
+    @State static var values: [Float] = [0.0, 0.1, 0.2, 0.3, 0.4, 1.0, 1.0, 0.5, 0.5, 0.0]
     static var previews: some View {
         Graph(values: $values)
     }
